requirements.txt
anyio==4.3.0
APScheduler==3.6.3
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
python-telegram-bot==20.8
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tornado==6.4
tzdata==2024.1
tzlocal==5.2
